Modification for checking the Stale data exception for Version data type.

+++ Commons/UICore/src/main/java/com/mfino/uicore/fix/processor/impl/ActorChannelMappingProcessorImpl.java
@@ -107,7 +107,7 @@ public class ActorChannelMappingProcessorImpl extends BaseFixProcessor implement
 			e.setGroupID(acm.getGroups().getId().toString());
 			e.setGroupName(acm.getGroups().getGroupname());
 		}
-		e.setRecordVersion(((Long)acm.getVersion()).intValue());
+		e.setRecordVersion(acm.getVersion());
 		e.setCreatedBy(acm.getCreatedby());
 		e.setCreateTime(acm.getCreatetime());
 		e.setUpdatedBy(acm.getUpdatedby());
